% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.clients.get}
\alias{accounts.clients.get}
\title{Gets a client buyer with a given client account ID.}
\usage{
accounts.clients.get(accountId, clientAccountId)
}
\arguments{
\item{accountId}{Numerical account ID of the client's sponsor buyer}

\item{clientAccountId}{Numerical account ID of the client buyer to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}
}

