% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.clients.list}
\alias{accounts.clients.list}
\title{Lists all the clients for the current sponsor buyer.}
\usage{
accounts.clients.list(accountId, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{accountId}{Unique numerical account ID of the sponsor buyer to list the clients for}

\item{pageSize}{Requested page size}

\item{pageToken}{A token identifying a page of results the server should return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}
}

