% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.clients.update}
\alias{accounts.clients.update}
\title{Updates an existing client buyer.}
\usage{
accounts.clients.update(Client, accountId, clientAccountId)
}
\arguments{
\item{Client}{The \link{Client} object to pass to this method}

\item{accountId}{Unique numerical account ID for the buyer of which the client buyer}

\item{clientAccountId}{Unique numerical account ID of the client to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}

Other Client functions: \code{\link{Client}},
  \code{\link{accounts.clients.create}}
}
