% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.creatives.create}
\alias{accounts.creatives.create}
\title{Creates a creative.}
\usage{
accounts.creatives.create(Creative, accountId, duplicateIdMode = NULL)
}
\arguments{
\item{Creative}{The \link{Creative} object to pass to this method}

\item{accountId}{The account that this creative belongs to}

\item{duplicateIdMode}{Indicates if multiple creatives can share an ID or not}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}

Other Creative functions: \code{\link{Creative}},
  \code{\link{accounts.creatives.update}}
}
