% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.creatives.get}
\alias{accounts.creatives.get}
\title{Gets a creative.}
\usage{
accounts.creatives.get(accountId, creativeId)
}
\arguments{
\item{accountId}{The account the creative belongs to}

\item{creativeId}{The ID of the creative to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}
}
