% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.creatives.list}
\alias{accounts.creatives.list}
\title{Lists creatives.}
\usage{
accounts.creatives.list(accountId, pageToken = NULL, pageSize = NULL,
  query = NULL)
}
\arguments{
\item{accountId}{The account to list the creatives from}

\item{pageToken}{A token identifying a page of results the server should return}

\item{pageSize}{Requested page size}

\item{query}{An optional query string to filter creatives}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}
}
