% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.creatives.stopWatching}
\alias{accounts.creatives.stopWatching}
\title{Stops watching a creative. Will stop push notifications being sent to thetopics when the creative changes status.}
\usage{
accounts.creatives.stopWatching(StopWatchingCreativeRequest, accountId,
  creativeId)
}
\arguments{
\item{StopWatchingCreativeRequest}{The \link{StopWatchingCreativeRequest} object to pass to this method}

\item{accountId}{The account of the creative to stop notifications for}

\item{creativeId}{The creative ID of the creative to stop notifications for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}

Other StopWatchingCreativeRequest functions: \code{\link{StopWatchingCreativeRequest}}
}
