% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\name{accounts.creatives.watch}
\alias{accounts.creatives.watch}
\title{Watches a creative. Will result in push notifications being sent to thetopic when the creative changes status.}
\usage{
accounts.creatives.watch(WatchCreativeRequest, accountId, creativeId)
}
\arguments{
\item{WatchCreativeRequest}{The \link{WatchCreativeRequest} object to pass to this method}

\item{accountId}{The account of the creative to watch}

\item{creativeId}{The creative ID to watch for status changes}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adexchange.buyer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/ad-exchange/buyer-rest/guides/client-access/}{Google Documentation}

Other WatchCreativeRequest functions: \code{\link{WatchCreativeRequest}}
}
