% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{AddDealAssociationRequest}
\alias{AddDealAssociationRequest}
\title{AddDealAssociationRequest Object}
\usage{
AddDealAssociationRequest(association = NULL)
}
\arguments{
\item{association}{The association between a creative and a deal that should be added}
}
\value{
AddDealAssociationRequest object
}
\description{
AddDealAssociationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request for associating a deal and a creative.
}
