% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{AppContext}
\alias{AppContext}
\title{AppContext Object}
\usage{
AppContext(appTypes = NULL)
}
\arguments{
\item{appTypes}{The app types this restriction applies to}
}
\value{
AppContext object
}
\description{
AppContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
