% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{AuctionContext}
\alias{AuctionContext}
\title{AuctionContext Object}
\usage{
AuctionContext(auctionTypes = NULL)
}
\arguments{
\item{auctionTypes}{The auction types this restriction applies to}
}
\value{
AuctionContext object
}
\description{
AuctionContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
