% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{Client}
\alias{Client}
\title{Client Object}
\usage{
Client(visibleToSeller = NULL, entityId = NULL, clientAccountId = NULL,
  entityName = NULL, status = NULL, entityType = NULL,
  clientName = NULL, role = NULL)
}
\arguments{
\item{visibleToSeller}{Whether the client buyer will be visible to sellers}

\item{entityId}{Numerical identifier of the client entity}

\item{clientAccountId}{The globally-unique numerical ID of the client}

\item{entityName}{The name of the entity}

\item{status}{The status of the client buyer}

\item{entityType}{The type of the client entity: `ADVERTISER`, `BRAND`, or `AGENCY`}

\item{clientName}{Name used to represent this client to publishers}

\item{role}{The role which is assigned to the client buyer}
}
\value{
Client object
}
\description{
Client Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A client resource represents a client buyer&mdash;an agency,a brand, or an advertiser customer of the sponsor buyer.Users associated with the client buyer have restricted access tothe Ad Exchange Marketplace and certain other sectionsof the Ad Exchange Buyer UI based on the rolegranted to the client buyer.All fields are required unless otherwise specified.
}
\seealso{
Other Client functions: \code{\link{accounts.clients.create}},
  \code{\link{accounts.clients.update}}
}
