% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ClientUser}
\alias{ClientUser}
\title{Ad Exchange Buyer API II Objects 
Accesses the latest features for managing Ad Exchange accounts and Real-Time Bidding configurations and auction metrics.}
\usage{
ClientUser(email = NULL, clientAccountId = NULL, status = NULL,
  userId = NULL)
}
\arguments{
\item{email}{User's email address}

\item{clientAccountId}{Numerical account ID of the client buyer}

\item{status}{The status of the client user}

\item{userId}{The unique numerical ID of the client user}
}
\value{
ClientUser object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:31:13
filename: /Users/mark/dev/R/autoGoogleAPI/googleadexchangebuyer2v2beta1.auto/R/adexchangebuyer2_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ClientUser Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A client user is created under a client buyer and has restricted access tothe Ad Exchange Marketplace and certain other sectionsof the Ad Exchange Buyer UI based on the rolegranted to the associated client buyer.The only way a new client user can be created is via accepting anemail invitation(see theaccounts.clients.invitations.createmethod).All fields are required unless otherwise specified.
}

