% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ClientUserInvitation}
\alias{ClientUserInvitation}
\title{ClientUserInvitation Object}
\usage{
ClientUserInvitation(email = NULL, clientAccountId = NULL,
  invitationId = NULL)
}
\arguments{
\item{email}{The email address to which the invitation is sent}

\item{clientAccountId}{Numerical account ID of the client buyer}

\item{invitationId}{The unique numerical ID of the invitation that is sent to the user}
}
\value{
ClientUserInvitation object
}
\description{
ClientUserInvitation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An invitation for a new client user to get access to the AdExchange Buyer UI.All fields are required unless otherwise specified.
}

