% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{Correction}
\alias{Correction}
\title{Correction Object}
\usage{
Correction(details = NULL, type = NULL, contexts = NULL)
}
\arguments{
\item{details}{Additional details about what was corrected}

\item{type}{The type of correction that was applied to the creative}

\item{contexts}{The contexts for the correction}
}
\value{
Correction object
}
\description{
Correction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
