% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{Creative}
\alias{Creative}
\title{Creative Object}
\usage{
Creative(detectedDomains = NULL, detectedAdvertiserIds = NULL,
  filteringStats = NULL, attributes = NULL, apiUpdateTime = NULL,
  detectedLanguages = NULL, creativeId = NULL, accountId = NULL,
  native = NULL, video = NULL, servingRestrictions = NULL,
  agencyId = NULL, clickThroughUrls = NULL,
  detectedSensitiveCategories = NULL, adChoicesDestinationUrl = NULL,
  restrictedCategories = NULL, corrections = NULL, version = NULL,
  vendorIds = NULL, impressionTrackingUrls = NULL, html = NULL,
  detectedProductCategories = NULL, dealsStatus = NULL,
  openAuctionStatus = NULL, advertiserName = NULL)
}
\arguments{
\item{detectedDomains}{@OutputOnly}

\item{detectedAdvertiserIds}{@OutputOnly Detected advertiser IDs, if any}

\item{filteringStats}{@OutputOnly The filtering stats for this creative}

\item{attributes}{All attributes for the ads that may be shown from this creative}

\item{apiUpdateTime}{@OutputOnly The last update timestamp of the creative via API}

\item{detectedLanguages}{@OutputOnly}

\item{creativeId}{The buyer-defined creative ID of this creative}

\item{accountId}{The account that this creative belongs to}

\item{native}{A native creative}

\item{video}{A video creative}

\item{servingRestrictions}{@OutputOnly The granular status of this ad in specific contexts}

\item{agencyId}{The agency ID for this creative}

\item{clickThroughUrls}{The set of destination URLs for the creative}

\item{detectedSensitiveCategories}{@OutputOnly Detected sensitive categories, if any}

\item{adChoicesDestinationUrl}{The link to AdChoices destination page}

\item{restrictedCategories}{All restricted categories for the ads that may be shown from this creative}

\item{corrections}{@OutputOnly Shows any corrections that were applied to this creative}

\item{version}{@OutputOnly The version of this creative}

\item{vendorIds}{All vendor IDs for the ads that may be shown from this creative}

\item{impressionTrackingUrls}{The set of URLs to be called to record an impression}

\item{html}{An HTML creative}

\item{detectedProductCategories}{@OutputOnly Detected product categories, if any}

\item{dealsStatus}{@OutputOnly The top-level deals status of this creative}

\item{openAuctionStatus}{@OutputOnly The top-level open auction status of this creative}

\item{advertiserName}{The name of the company being advertised in the creative}
}
\value{
Creative object
}
\description{
Creative Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A creative and its classification data.
}
\seealso{
Other Creative functions: \code{\link{accounts.creatives.create}},
  \code{\link{accounts.creatives.update}}
}
