% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{CreativeDealAssociation}
\alias{CreativeDealAssociation}
\title{CreativeDealAssociation Object}
\usage{
CreativeDealAssociation(accountId = NULL, creativeId = NULL,
  dealsId = NULL)
}
\arguments{
\item{accountId}{The account the creative belongs to}

\item{creativeId}{The ID of the creative associated with the deal}

\item{dealsId}{The externalDealId for the deal associated with the creative}
}
\value{
CreativeDealAssociation object
}
\description{
CreativeDealAssociation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The association between a creative and a deal.
}
