% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{Date}
\alias{Date}
\title{Ad Exchange Buyer API II Objects 
Accesses the latest features for managing Ad Exchange accounts, Real-Time Bidding configurations and auction metrics, and Marketplace programmatic deals.}
\usage{
Date(year = NULL, day = NULL, month = NULL)
}
\arguments{
\item{year}{Year of date}

\item{day}{Day of month}

\item{month}{Month of year}
}
\value{
Date object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:20:56
filename: /Users/mark/dev/R/autoGoogleAPI/googleadexchangebuyer2v2beta1.auto/R/adexchangebuyer2_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Date Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a whole calendar date, e.g. date of birth. The time of day andtime zone are either specified elsewhere or are not significant. The dateis relative to the Proleptic Gregorian Calendar. The day may be 0 torepresent a year and month where the day is not significant, e.g. credit cardexpiration date. The year may be 0 to represent a month and day independentof year, e.g. anniversary date. Related types are google.type.TimeOfDayand `google.protobuf.Timestamp`.
}
