% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{Disapproval}
\alias{Disapproval}
\title{Disapproval Object}
\usage{
Disapproval(details = NULL, reason = NULL)
}
\arguments{
\item{details}{Additional details about the reason for disapproval}

\item{reason}{The categorized reason for disapproval}
}
\value{
Disapproval object
}
\description{
Disapproval Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
