% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{FilteringStats}
\alias{FilteringStats}
\title{FilteringStats Object}
\usage{
FilteringStats(reasons = NULL, date = NULL)
}
\arguments{
\item{reasons}{The set of filtering reasons for this date}

\item{date}{The day during which the data was collected}
}
\value{
FilteringStats object
}
\description{
FilteringStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
