% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{HtmlContent}
\alias{HtmlContent}
\title{HtmlContent Object}
\usage{
HtmlContent(height = NULL, width = NULL, snippet = NULL)
}
\arguments{
\item{height}{The height of the HTML snippet in pixels}

\item{width}{The width of the HTML snippet in pixels}

\item{snippet}{The HTML snippet that displays the ad when inserted in the web page}
}
\value{
HtmlContent object
}
\description{
HtmlContent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
HTML content for a creative.
}
