% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(width = NULL, url = NULL, height = NULL)
}
\arguments{
\item{width}{Image width in pixels}

\item{url}{The URL of the image}

\item{height}{Image height in pixels}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An image resource. You may provide a larger image than was requested,so long as the aspect ratio is preserved.
}
