% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ListClientsResponse}
\alias{ListClientsResponse}
\title{ListClientsResponse Object}
\usage{
ListClientsResponse(nextPageToken = NULL, clients = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{clients}{The returned list of clients}
}
\value{
ListClientsResponse object
}
\description{
ListClientsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

