% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ListClientUserInvitationsResponse}
\alias{ListClientUserInvitationsResponse}
\title{ListClientUserInvitationsResponse Object}
\usage{
ListClientUserInvitationsResponse(nextPageToken = NULL, invitations = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{invitations}{The returned list of client users}
}
\value{
ListClientUserInvitationsResponse object
}
\description{
ListClientUserInvitationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
