% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ListClientUsersResponse}
\alias{ListClientUsersResponse}
\title{ListClientUsersResponse Object}
\usage{
ListClientUsersResponse(nextPageToken = NULL, users = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{users}{The returned list of client users}
}
\value{
ListClientUsersResponse object
}
\description{
ListClientUsersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
