% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ListCreativesResponse}
\alias{ListCreativesResponse}
\title{ListCreativesResponse Object}
\usage{
ListCreativesResponse(nextPageToken = NULL, creatives = NULL)
}
\arguments{
\item{nextPageToken}{A token to retrieve the next page of results}

\item{creatives}{The list of creatives}
}
\value{
ListCreativesResponse object
}
\description{
ListCreativesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response for listing creatives.
}
