% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ListDealAssociationsResponse}
\alias{ListDealAssociationsResponse}
\title{ListDealAssociationsResponse Object}
\usage{
ListDealAssociationsResponse(associations = NULL, nextPageToken = NULL)
}
\arguments{
\item{associations}{The list of associations}

\item{nextPageToken}{A token to retrieve the next page of results}
}
\value{
ListDealAssociationsResponse object
}
\description{
ListDealAssociationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response for listing creative and deal associations
}
