% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{LocationContext}
\alias{LocationContext}
\title{LocationContext Object}
\usage{
LocationContext(geoCriteriaIds = NULL)
}
\arguments{
\item{geoCriteriaIds}{IDs representing the geo location for this context}
}
\value{
LocationContext object
}
\description{
LocationContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
