% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{NativeContent}
\alias{NativeContent}
\title{NativeContent Object}
\usage{
NativeContent(storeUrl = NULL, headline = NULL, appIcon = NULL,
  callToAction = NULL, body = NULL, starRating = NULL, videoUrl = NULL,
  logo = NULL, clickLinkUrl = NULL, priceDisplayText = NULL,
  image = NULL, clickTrackingUrl = NULL, advertiserName = NULL)
}
\arguments{
\item{storeUrl}{The URL to the app store to purchase/download the promoted app}

\item{headline}{A short title for the ad}

\item{appIcon}{The app icon, for app download ads}

\item{callToAction}{A label for the button that the user is supposed to click}

\item{body}{A long description of the ad}

\item{starRating}{The app rating in the app store}

\item{videoUrl}{The URL to fetch a native video ad}

\item{logo}{A smaller image, for the advertiser's logo}

\item{clickLinkUrl}{The URL that the browser/SDK will load when the user clicks the ad}

\item{priceDisplayText}{The price of the promoted app including currency info}

\item{image}{A large image}

\item{clickTrackingUrl}{The URL to use for click tracking}

\item{advertiserName}{The name of the advertiser or sponsor, to be displayed in the ad creative}
}
\value{
NativeContent object
}
\description{
NativeContent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Native content for a creative.
}
