% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{PlatformContext}
\alias{PlatformContext}
\title{PlatformContext Object}
\usage{
PlatformContext(platforms = NULL)
}
\arguments{
\item{platforms}{The platforms this restriction applies to}
}
\value{
PlatformContext object
}
\description{
PlatformContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
