% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{Reason}
\alias{Reason}
\title{Reason Object}
\usage{
Reason(count = NULL, status = NULL)
}
\arguments{
\item{count}{The number of times the creative was filtered for the status}

\item{status}{The filtering status code}
}
\value{
Reason object
}
\description{
Reason Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A specific filtering status and how many times it occurred.
}
