% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{SecurityContext}
\alias{SecurityContext}
\title{SecurityContext Object}
\usage{
SecurityContext(securities = NULL)
}
\arguments{
\item{securities}{The security types in this context}
}
\value{
SecurityContext object
}
\description{
SecurityContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
