% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ServingContext}
\alias{ServingContext}
\title{ServingContext Object}
\usage{
ServingContext(all = NULL, appType = NULL, securityType = NULL,
  platform = NULL, location = NULL, auctionType = NULL)
}
\arguments{
\item{all}{Matches all contexts}

\item{appType}{Matches impressions for a particular app type}

\item{securityType}{Matches impressions for a particular security type}

\item{platform}{Matches impressions coming from a particular platform}

\item{location}{Matches impressions coming from users *or* publishers in a specific}

\item{auctionType}{Matches impressions for a particular auction type}
}
\value{
ServingContext object
}
\description{
ServingContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The serving context for this restriction.
}
