% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ServingRestriction}
\alias{ServingRestriction}
\title{ServingRestriction Object}
\usage{
ServingRestriction(contexts = NULL, status = NULL,
  disapprovalReasons = NULL)
}
\arguments{
\item{contexts}{The contexts for the restriction}

\item{status}{The status of the creative in this context (for example, it has been}

\item{disapprovalReasons}{Any disapprovals bound to this restriction}
}
\value{
ServingRestriction object
}
\description{
ServingRestriction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
