% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{StopWatchingCreativeRequest}
\alias{StopWatchingCreativeRequest}
\title{StopWatchingCreativeRequest Object}
\usage{
StopWatchingCreativeRequest()
}
\value{
StopWatchingCreativeRequest object
}
\description{
StopWatchingCreativeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request for stopping notifications for changes to creative Status.
}
\seealso{
Other StopWatchingCreativeRequest functions: \code{\link{accounts.creatives.stopWatching}}
}
