% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{VideoContent}
\alias{VideoContent}
\title{VideoContent Object}
\usage{
VideoContent(videoUrl = NULL)
}
\arguments{
\item{videoUrl}{The URL to fetch a video ad}
}
\value{
VideoContent object
}
\description{
VideoContent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Video content for a creative.
}
