% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{WatchCreativeRequest}
\alias{WatchCreativeRequest}
\title{WatchCreativeRequest Object}
\usage{
WatchCreativeRequest(topic = NULL)
}
\arguments{
\item{topic}{The Pub/Sub topic to publish notifications to}
}
\value{
WatchCreativeRequest object
}
\description{
WatchCreativeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request for watching changes to creative Status.
}
\seealso{
Other WatchCreativeRequest functions: \code{\link{accounts.creatives.watch}}
}
