% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_functions.R
\docType{package}
\name{adexchangebuyer2_googleAuthR}
\alias{adexchangebuyer2_googleAuthR}
\alias{adexchangebuyer2_googleAuthR-package}
\title{Ad Exchange Buyer API II
Accesses the latest features for managing Ad Exchange accounts and Real-Time Bidding configurations and auction metrics.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:31:13
filename: /Users/mark/dev/R/autoGoogleAPI/googleadexchangebuyer2v2beta1.auto/R/adexchangebuyer2_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.buyer
}
}

