% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangebuyer2_objects.R
\name{ClientUser}
\alias{ClientUser}
\title{ClientUser Object}
\usage{
ClientUser(status = NULL, userId = NULL, email = NULL,
  clientAccountId = NULL)
}
\arguments{
\item{status}{The status of the client user}

\item{userId}{The unique numerical ID of the client user}

\item{email}{User's email address}

\item{clientAccountId}{Numerical account ID of the client buyer}
}
\value{
ClientUser object
}
\description{
ClientUser Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A client user is created under a client buyer and has restricted access tothe Ad Exchange Marketplace and certain other sectionsof the Ad Exchange Buyer UI based on the rolegranted to the associated client buyer.The only way a new client user can be created is via accepting anemail invitation(see theaccounts.clients.invitations.createmethod).All fields are required unless otherwise specified.
}
